fix: address PR review feedback
- Fix logger reference to removed polish.enabled field
- Add tapPolishEnabled/holdPolishEnabled to DEFAULT_SHORTCUT
- Extract DEFAULT_TAP_POLISH_ENABLED and DEFAULT_HOLD_POLISH_ENABLED constants
- Share PolishConfig interface from shared/app-state.ts
- Remove duplicate PolishConfig definitions from electron/config and UI components
